PyTorch/computer_vision/segmentation/Unet/requirements_u22.txt
git+https://github.com/NVIDIA/dllogger
nibabel==3.1.1
scikit-learn==1.2.1
pynvml==8.0.4
pytorch_ranger==0.1.1
dropblock==0.3.0
monai==0.4.0
torch-optimizer==0.0.1a15
scikit-image==0.19.3
awscli
lightning==2.0.6
lightning-habana==1.0.1
numpy==1.23.1
